Break out JoinsValuesBuilder into Helper folder
The logic will be reused with sorting.
